Adding null process test suite.
Created a suite of tests which all block in succession so that we are
stuck permanently running the null process to ensure there are no
problems as a result of that.  Added a carriage return before any output
so that that irritating space at the start of output isn't seen.